[to server]

+++ src/cl/fpga/simple_addition.cl
@@ -7,10 +7,10 @@
 
 #define GROUP_SIZE 256
 
-__kernel void aplusb(__global float* restrict a,
-                     __global float* restrict b,
-                     __global float* restrict c,
-                     unsigned int n)
+__kernel void aplusb(__global uint* restrict a,
+                     __global uint* restrict b,
+                     __global uint* restrict c,
+                    uint n)
 {
     const unsigned int index = get_global_id(0);
     c[index] = a[index] + b[index];
